Enum gluster::VolumeType

pub enum VolumeType {
    Distribute,
    Stripe,
    Replicate,
    StripedAndReplicate,
    Disperse,
    DistributedAndStripe,
    DistributedAndReplicate,
    DistributedAndStripedAndReplicate,
    DistributedAndDisperse,
}

These are all the different Volume types that are possible in Gluster Note: Tier is not represented here because I'm waiting for it to become more stable For more information about these types see: Gluster Volume

Variants

Distribute

Stripe

Replicate

StripedAndReplicate

Disperse

DistributedAndStripe

DistributedAndReplicate

DistributedAndStripedAndReplicate

DistributedAndDisperse

Methods

impl VolumeType

fn new(name: &str) -> VolumeType

Constructs a new VolumeType from a &str

fn to_string(self) -> String

Returns a String representation of the selected enum variant.